+++ src/main/java/NeighbourT.java
@@ -2,7 +2,7 @@ package main.java;
 
 class NeighbourT{
     int target; /* the index of cluster node representing neighbour */
-    float distance; /* distance between the nodes */
+    double distance; /* distance between the nodes */
     NeighbourT next; /* linked list entries */
     NeighbourT prev; /* linked list entries */
 }
